% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplot.R
\name{dataplot}
\alias{dataplot}
\title{base function to build build up graphs element by element}
\usage{
dataplot(data_frame, x_var, y_var, element_function, main_title = NULL,
  bar_width = NULL, x_title = ggplot2::waiver(), x_ticks = TRUE,
  y_title = ggplot2::waiver(), x_limits = NULL, y_limits = NULL,
  x_breaks = ggplot2::waiver(), y_breaks = ggplot2::waiver(),
  y_facet = FALSE, label_vars = FALSE, percentage = FALSE,
  colour_vars = NULL, fill_vars = NULL, colour_palette = NULL,
  fill_palette = NULL, pos = "dodge", barplot_alpha = 1,
  colour_title = FALSE, fill_title = FALSE,
  legend_labels = ggplot2::waiver(), legend_rev = FALSE, flip = FALSE)
}
\description{
base function to build build up graphs element by element
}
