% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bar_theme.R
\name{set_bar_theme}
\alias{set_bar_theme}
\title{Sets a theme for a barplot}
\usage{
set_bar_theme(text_col = black_seq[8], basesize = 12,
  legend_col = "#E5E2E0", legend_pos = "bottom", keysize = 0.7)
}
\arguments{
\item{text_col}{Text colour, defaults to black_seq[8]}

\item{basesize}{Text basesize, defaults to 12}

\item{legend_col}{legend background colour, default at "#E5E2E0"}

\item{legend_pos}{legend positions, takes a string ("top", "bottom", 
"right", "left"), or a vector of two numbers between 0 and 1, indicating 
the relative x and y positions.}

\item{keysize}{size of the legend}
}
\description{
Sets a theme for a barplot
}
